Refactoring backstage
Added isUrl back in
